Added images to recipe pages
I also resized the turkey-burgers image as it was very large to start.
I set the width to 200 and allowed the height to scale accordingly.
I included the index.html here which was a mistake. I was worried
that I forgot a change but should have just checked git status.